update classes, methods related to function: "update informations of project, including rooms and its images"

+++ src/main/java/com/interchange/service/impl/ImageRoomServiceImpl.java
@@ -1,10 +1,12 @@
 package com.interchange.service.impl;
 
 import com.interchange.base.BaseResponse;
+import com.interchange.entities.ImageBlog;
 import com.interchange.entities.ImageRoom;
 import com.interchange.entities.Project;
 import com.interchange.entities.Room;
 import com.interchange.repository.ImageRoomRepository;
+import com.interchange.repository.ProjectRepository;
 import com.interchange.repository.RoomRepository;
 import com.interchange.service.ImageRoomService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -37,12 +39,17 @@ public class ImageRoomServiceImpl extends BaseResponse  implements ImageRoomServ
         for (int i = 0; i < multipartFiles.length; i++) {
             ImageRoom imageRoom = convertFile(multipartFiles[i]);
             Room room = project.getRooms().get(i);
-            int fileId = room.getImageRoom().getFileId();
-            room.setImageRoom(null);
-            imageRoomRepository.deleteById(fileId);
-            roomRepository.flush();
-            imageRoomRepository.flush();
-            imageRoom.setRoom(room);
+            if(room.getImageRoom() != null) {
+                int fileId = room.getImageRoom().getFileId();
+                room.setImageRoom(null);
+                imageRoomRepository.deleteById(fileId);
+                roomRepository.flush();
+                imageRoomRepository.flush();
+                imageRoom.setRoom(room);
+            }
+            else {
+                imageRoom.setRoom(room);
+            }
             imageRoomRepository.save(imageRoom);
         }
         return null;
@@ -51,6 +58,12 @@ public class ImageRoomServiceImpl extends BaseResponse  implements ImageRoomServ
     @Override
     public ResponseEntity<?> getImageRoomByRoomId(int roomId) {
         return getResponseEntityRoomFile(roomRepository.findById(roomId).get().getImageRoom());
+    }
+
+    @Override
+    public ResponseEntity<?> isRoomHasImage(int roomId) {
+        boolean flag = imageRoomRepository.findByRoom_RoomId(roomId) != null;
+        return getResponseEntity(flag);
     }
 
 }
